# Add Spark cast expr based on cast hooks