--- a/velox/expression/CastExpr.h
+++ b/velox/expression/CastExpr.h
@@ -16,6 +16,7 @@
 
 #pragma once
 
+#include "velox/expression/CastHooks.h"
 #include "velox/expression/FunctionCallToSpecialForm.h"
 #include "velox/expression/SpecialForm.h"
 
@@ -72,14 +73,20 @@ class CastExpr : public SpecialForm {
   /// @param type The target type of the cast expression
   /// @param expr The expression to cast
   /// @param trackCpuUsage Whether to track CPU usage
-  CastExpr(TypePtr type, ExprPtr&& expr, bool trackCpuUsage, bool nullOnFailure)
+  CastExpr(
+      TypePtr type,
+      ExprPtr&& expr,
+      bool trackCpuUsage,
+      bool nullOnFailure,
+      std::shared_ptr<CastHooks> hooks)
       : SpecialForm(
             type,
             std::vector<ExprPtr>({expr}),
             nullOnFailure ? kTryCast.data() : kCast.data(),
             false /* supportsFlatNoNullsFastPath */,
             trackCpuUsage),
-        nullOnFailure_(nullOnFailure) {}
+        nullOnFailure_(nullOnFailure),
+        hooks_(std::move(hooks)) {}
 
   void evalSpecialForm(
       const SelectivityVector& rows,
@@ -273,6 +280,7 @@ class CastExpr : public SpecialForm {
   folly::F14FastMap<std::string, CastOperatorPtr> castOperators_;
 
   bool nullOnFailure_;
+  std::shared_ptr<CastHooks> hooks_;
 
   bool inTopLevel = false;
 };