… velox/docs/functions/presto/conversion.rst
… velox/docs/functions/spark/conversion.rst
… velox/expression/CMakeLists.txt
… velox/expression/tests/CastExprTest.cpp
… velox/functions/prestosql/tests/CastBaseTest.h
… velox/functions/sparksql/specialforms/CMakeLists.txt
… velox/functions/sparksql/tests/CMakeLists.txt
… velox/functions/sparksql/tests/SparkCastExprTest.cpp