There are several corner cases of semantic differences on cast between Presto 
and Spark. According to the implementation of
`registerFunctionCallToSpecialForm`, a previously registered special form can be 
overriden by one registered after it and of the same name.This PR implements 
SparkCastExpr, which is registered as special form, and can be customized with 
the help of cast hooks compatible with Spark's semantics. Below hooks are added 
to solve several semantic differences.
- castStringToTimestamp
- castStringToDate
- castTimestampToString
- legacy
- removeWhiteSpaces
- truncate

Two configurations `kCastToIntByTruncate` and `kCastStringToDateIsIso8601` are 
replaced by cast hooks. These configurations are no longer used and will be removed 
by subsequent PRs.

https://github.com/facebookincubator/velox/issues/4876
Fixes https://github.com/facebookincubator/velox/issues/8121.