@rui-mo 2023-11-03
@mbasmanova Thank you for the review. Can I prototype with `kCastStringToDateIsIso8601` in this PR? That is to move the Spark-specific behavior controlled by this config into SparkCastExpr, and remove this config.