@rui-mo 2023-11-07
@mbasmanova Masha, I made the prototype with `IsIso8601`. Could you help review? Thanks.

> Overriding CAST requires also customizing this code in ExprCompiler.

Seems it is not necessary if we don't want to parse cast from string. In this PR, SparkCastExpr can to be created with `CallTypedExpr`, but if we want to parse from string like `cast (c0 as int)`, more changes are required. Besides what is included in f8eb9237f128b81524a8fe1663e6dc38149a4c6c, we need to decide whether to create CastExpr or SparkCastExpr in [parseCastExpr](https://github.com/facebookincubator/velox/blob/main/velox/duckdb/conversion/DuckParser.cpp#L551-L552) (maybe through a parse option).

In this PR, parsing from string to SparkCastExpr is not included. But if it is needed, I can continue to work on that. Seems it is also a UT conveniency issue.