@rui-mo 2023-12-04
@mbasmanova Thank you for the suggestions. Fixed them, and apology for the delay.
> Figure out all the differences between Spark and Presto CAST impementations to see whether it is feasible to continue using CastExpr as a base class.

Summarized them in this table: https://github.com/facebookincubator/velox/issues/4876#issuecomment-1838011064.