@mbasmanova 2023-12-04
> Summarized them in this table: https://github.com/facebookincubator/velox/issues/4876#issuecomment-1838011064.

@rui-mo Thank you, Rui. What's your assessment of these differences? Is it feasible to continue using CastExpr as a base class?