@rui-mo 2023-12-05
> What's your assessment of these differences? Is it feasible to continue using CastExpr as a base class?

@mbasmanova These cases can be divided into below categories.

1. `cast_to_int_by_truncate` behaviors.
2. overflow when casting to integral types.
3. cast from string to other type.
4. cast from other type to string.

As discussed with @PHILO-HE, we tend to use CastExpr as a base class, which can help reuse existing code and avoid duplicates. Take below case as an example.
| Conversion |  Cases | Presto result | Spark result |
| ------------- | ------------- |  ------------- | ------------ |
| string -> integral/floating-point/decimal/date/timestamp | SELECT cast('\t\n\u001F 123\u000B' as int); | invalid | 123 |

we can fix some of them by overriding a `getCastFromStringOutput` method of the base class, and removes the white-spaces (which is required in Spark for cast from string) before calling the cast function.
https://github.com/facebookincubator/velox/blob/261d0b0a37606355503fe218cbd240f764cc9f05/velox/expression/CastExpr-inl.h#L232-L233
 
As for https://github.com/facebookincubator/velox/pull/5307#discussion_r1307094037, we don't need to implement duplicate cast functions in CastExpr and SparkCastExpr. By overriding a small part of code in SparkCastExpr to remove the white-spaces before cast should be enough.

Glad to hear your opinions, thanks.