@rui-mo 2023-12-15
> Have you had a chance to check for performance regressions. Do you see any?

@mbasmanova I checked below three benchmarks because their hooks bring extra if-else clause besides virtual function overhead. Below is the benchmark results I got. Looks `getDecimalIntegral` brings some overhead. What do you think?

| benchmark  | related hook | before this PR | after this PR |
| ------------- | ------------- | ------------- | ------------- |
| cast##cast_timestamp  | castTimestampToString | 60.76ms, 16.46 iters/s   | 60.40ms, 16.56 iters/s  |
| cast##cast_real_as_int  | truncate | 7.66ms, 130.52 iters/s  | 8.03ms, 124.60 iters/s  |
| cast##cast_decimal_as_bigint  | getDecimalIntegral | 5.05ms, 198.16 iters/s  | 8.08ms, 123.78 iters/s  |
```