@mbasmanova 2023-12-15
@rui-mo I see that cast_decimal_as_bigint got slower. Have you had a chance to profile to see why? CC: @kagamiori @laithsakka @gggrace14