@rui-mo 2023-12-15
@mbasmanova Yes, I will check the effect to `cast_decimal_as_bigint`, thanks.