@rui-mo 2023-12-18
@mbasmanova I find the performance regression comes from `getDecimalIntegral` hook. Removed it with https://github.com/facebookincubator/velox/pull/7377/commits/8c4fd89d91b7083c5eab1dc9de19fb6f322be307 and use truncate flag instead as https://github.com/facebookincubator/velox/pull/7377/files#diff-872249bbe84c54f0119d4568444009b9b850ec34b802ebbc0b641bb012fa02baR344-R357. The performance is back.
> cast##cast_decimal_as_bigint                                4.68ms    213.51 iters/s