@kagamiori 2024-01-09
Hi @rui-mo, the removal of QueryConfig:: kCastToIntByTruncate and QueryConfig:: kCastStringToDateIsIso8601 causes some internal test failures. It turns out that presto/presto-native-execution in the Presto repo is relying on kCastToIntByTruncate. Changes in presto/presto-native-execution must be made on GitHub first, so I couldn't resolve the issue internally for you. Could you help do the following the resolve the issue?

1. Update this PR to not remove the QueryConfigs. The rest stays unchanged. This would avoid breaking tests that rely on presto/presto-native-execution.
2. Make a PR in the Presto repo to remove the dependency on the query config (https://github.com/prestodb/presto/blob/9ca2ea67c0a25fb3e20799c76a38cfbad4f8d663/presto-native-execution/presto_cpp/main/common/Configs.cpp#L683-L684).
3. After the Presto PR is merged, make another Velox PR to remove QueryConfig:: kCastToIntByTruncate.

Please let me know if you have any questions or thoughts. Thanks!