@rui-mo 2024-01-10
@kagamiori Thanks for providing the detailed steps, and they make sense to me. In this PR, the two configs are added back, but they are not controlling any execution logic now. Opened https://github.com/prestodb/presto/pull/21663 in Presto to remove the dependency. It seems `kCastStringToDateIsIso8601` is not in use of presto_cpp. Could you help confirm? Thank you.