@kagamiori 2024-01-10
FYI, I replace
```
static void SetUpTestCase() {
    functions::sparksql::registerFunctions("");
  }
``` 
in the `SparkCastExprTest` class with
```
SparkCastExprTest() {
    functions::sparksql::registerFunctions("");
  }
```
to resolve some internal test failures.