@kagamiori 2024-01-11
@rui-mo, this change turns out to break some string-related unit test in spark, and hence the linux build job in CI (https://app.circleci.com/pipelines/github/facebookincubator/velox/42248/workflows/18d750f7-eed8-466f-94d8-63958efad94a/jobs/291869/tests), so we're going to revert it. Could you take a look at the failed unit tests?