@rui-mo 2024-01-11
@kagamiori Thank you for helping land this PR. I know the cause of this failure, and please check https://github.com/facebookincubator/velox/issues/8121. Could you tell me more details about the internal failures as https://github.com/facebookincubator/velox/pull/7377#issuecomment-1885308164 mentioned?