@kagamiori 2024-01-11
> @kagamiori Thank you for helping land this PR. I know the cause of this failure, and please check #8121. Could you tell me more details about the internal failures as [#7377 (comment)](https://github.com/facebookincubator/velox/pull/7377#issuecomment-1885308164) mentioned?

Hi @rui-mo, we're seeing the CI signal "ci/circleci: linux-build" started failing on the main branch since #7377 was landed (https://github.com/facebookincubator/velox/commits/main/).